backend/requirements.txt
# Core
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.3.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# MikroTik
routeros-api  # use latest available on PyPI; pin to a specific repo/version if needed
paramiko==3.5.0
pyotp==2.9.0

# Async & Queue
celery==5.3.4
eventlet==0.33.3

# Monitoring
prometheus-client==0.18.0
prometheus-flask-exporter==0.22.4

# Time Series
influxdb-client[ciso]==1.40.0

# AI & Chat
openai==0.28.0

# Notifications
twilio==8.10.0
pywhatkit==5.4

# PDF & Documents
reportlab==4.0.4
qrcode==7.4.2
Pillow==10.0.0

# Security
bcrypt==4.0.1
cryptography==41.0.5

# Utils
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2

# Production
gunicorn==21.2.0
gevent==23.9.1

Flask-Caching==1.11.1

# Development & Linting
ruff==0.1.9
pytest==8.3.5
